fix: the incorrect split of byoc commands and args.
